+++ RELEASE_NOTES.md
@@ -6,3 +6,4 @@
 
 - The installation script uses `/bin/sh` instead `/bin/bash`
 - The release process creates ARM binaries
+- By default, install the Stevedore binary in $HOME/bin